--- test/__init__.py
@@ -1,16 +1,0 @@
-from .test_irf import TestPvoigtIRF
-from .test_exp_conv_irf import TestExpConvIRF
-from .test_dmp_osc_conv_irf import TestDmpOscConvIRF
-from .test_grad_exp_conv_irf import TestDerivExpConvIRF
-from .test_grad_dmp_osc_conv_irf import TestDerivDmpOscConvIRF
-from .test_grad_res_transient import TestGradResTransient
-from .test_grad_res_static import TestGradResStatic
-from .test_grad_peak_shape import TestDerivPeakShape
-from .test_fact_anal import TestFactAnal
-from .test_rate_eq import TestRateEqSolver
-from .test_ampgo import TestAMPGO
-
-__all__ = ['TestPvoigtIRF', 'TestExpConvIRF', 'TestDmpOscConvIRF',
-'TestDerivExpConvIRF', 'TestDerivDmpOscConvIRF', 'TestDerivPeakShape',
-'TestGradResTransient', 'TestGradResStatic',
-'TestFactAnal', 'TestRateEqSolver', 'TestAMPGO']
